feat(playground): add navigation using App layout
