+++ playground/package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "primeicons": "^7.0.0",
     "primevue": "^4.3.7",
-    "vue": "^3.5.18"
+    "vue": "^3.5.18",
+    "vue-router": "^4.3.0"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^6.0.1",
