+++ playground/src/components/RouterLink.vue
@@ -1,0 +1,17 @@
+<script setup>
+import { RouterLink } from 'vue-router';
+import { useAttrs } from 'vue';
+
+const props = defineProps({
+  href: {
+    type: String,
+    required: true,
+  },
+});
+
+const attrs = useAttrs();
+</script>
+
+<template>
+  <RouterLink :to="props.href" v-bind="attrs"><slot /></RouterLink>
+</template>
