+++ playground/src/main.js
@@ -4,11 +4,13 @@ import StyleClass from 'primevue/styleclass';
 import Tooltip from 'primevue/tooltip';
 import ToastService from 'primevue/toastservice';
 import App from './App.vue';
+import router from './router';
 import 'primeicons/primeicons.css';
 import './style.css';
 
 // Create and mount the Vue application
 const app = createApp(App);
+app.use(router);
 app.use(ToastService);
 app.use(PrimeVue, { unstyled: true });
 app.directive('tooltip', Tooltip);
