+++ playground/src/pages/Home.vue
@@ -1,0 +1,6 @@
+<template>
+  <section class="p-4 space-y-2">
+    <h1 class="text-2xl font-bold">Atlas UI Playground</h1>
+    <p>Select a page from the navigation to preview components.</p>
+  </section>
+</template>
